+++ plugins/modules/na_ontap_kerberos_realm.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 '''
 (c) 2019, Red Hat, Inc
-(c) 2019-2025, NetApp, Inc
+(c) 2019-2026, NetApp, Inc
 GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 '''
 
@@ -114,9 +114,32 @@ options:
     type: str
     version_added: '20.4.0'
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
 notes:
   - supports ZAPI and REST. REST requires ONTAP 9.6 or later.
   - supports check mode.
+  - Supports AWS Lambda proxy functionality when using REST. See README for example usage.
 '''
 
 EXAMPLES = '''
@@ -195,11 +218,13 @@ class NetAppOntapKerberosRealm:
             ad_server_ip=dict(required=False, type='str'),
             ad_server_name=dict(required=False, type='str')
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
 
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
             required_if=[
+                ('use_lambda', True, ['lambda_config']),
                 ('state', 'present', ['kdc_vendor', 'kdc_ip']),
                 ('kdc_vendor', 'microsoft', ['ad_server_ip', 'ad_server_name'])
             ]
@@ -214,6 +239,8 @@ class NetAppOntapKerberosRealm:
         self.svm_uuid = None
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])