+++ plugins/modules/na_ontap_publickey.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2021-2025, NetApp, Inc
+# (c) 2021-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 '''
@@ -62,10 +62,32 @@ options:
       - The name of the vserver to use.
       - Omit this option for cluster scoped user accounts.
     type: str
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
 
 notes:
   - This module supports check_mode.
   - This module is not idempotent if index is omitted.
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 '''
 
 EXAMPLES = """
@@ -139,11 +161,15 @@ class NetAppOntapPublicKey:
             public_key=dict(type='str'),
             vserver=dict(type='str'),
         ))
+        argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
 
         self.module = AnsibleModule(
             argument_spec=argument_spec,
             mutually_exclusive=[
                 ('delete_all', 'index')
+            ],
+            required_if=[
+                ('use_lambda', True, ['lambda_config'])
             ],
             supports_check_mode=True
         )